requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.8
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
sqlparse==0.5.1
tzdata==2024.1
